использование общего app_bar в группе для пользователя-не участника
